Reorganized utilites
