+++ src/utilities/UtilsGraphics.java
@@ -1,4 +1,4 @@
-package managers;
+package utilities;
 
 import java.awt.*;
 
@@ -6,7 +6,7 @@ import java.awt.*;
  * Created by Aunmag on 2017.06.12.
  */
 
-public class Utils {
+public class UtilsGraphics {
 
     private static final AlphaComposite alphaDefault = AlphaComposite.getInstance(
             AlphaComposite.SRC_OVER, 1
